feat[submission]: install shadcn for UI, fix folder structure
